+++ django_boost/template/__init__.py
@@ -1,0 +1,3 @@
+from django_boost.template.base import StrictInvalidTemplateVariable
+
+__all__ = ("StrictInvalidTemplateVariable",)